+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.0-alpha+4
+* Made `injections` in `RequestContext` private.
+* Renamed `properties` in `AngelBase` to `configuration`.
+
 # 1.1.0-alpha+3
 * Fixed a bug where `encoders` would cause a malformed response to be sent.
 * Fixed a bug where `encoders` would not always use the correct encoder.